perf(bin): optimistically send datagram, only await on failure
Previously `neqo-bin` would poll a UDP socket to be writable before
attempting to send a datagram. This results in 2 syscalls per send.

Instead, optimistically attempt to send the datagram first, on failure
`WouldBlock`, await and then try again. Under the assumption that the
majority of sends work on the first try, this reduces the number of
syscalls.

Follow-up to https://phabricator.services.mozilla.com/D239162#8315719.