fix: Mark all packets TX'ed before PTO as lost (#2129)
* fix: Mark all packets TX'ed before PTO as lost

We'd previously only mark 1 one or two packets as lost when a PTO fired.
That meant that we potentially didn't RTX all data that we could have
(i.e., that was in lost packets that we didn't mark lost).

This also changes the probing code to suppress redundant keep-alives,
i.e., PINGs that we sent for other reasons, which could double as
keep-alives but did not.

Broken out of #1998

* Fix merge

* fmt

* docs: remove outdated PTO timer reference

* test(transport): expect PTO to retransmit frames across datagrams

On PTO, expect endpoint to retransmit frames across many previously sent
datagrams, instead of just the last 1 or 2 datagrams.

* fix: don't consider all sent packets for retransmission on PTO

Follow-up to discussion https://github.com/mozilla/neqo/pull/2129#discussion_r1775388764.

---------

Signed-off-by: Lars Eggert <lars@eggert.org>
Co-authored-by: Max Leonard Inden <mail@max-inden.de>