fix(bin/bench): on upload, do not also request large download
The `neqo-bin` Download benchmark has the client do a single HTTP GET to
the server, requesting the number of bytes to download encoded in the
URL path, i.e. `http://[::1]:12345/104857600`.

The `neqo-bin` Upload benchmark has the client do a single HTTP POST to
the server, sending `104857600` random bytes along with it. That said,
previously it would also request the same amount of bytes to download
from the server.

This explains the Download / Upload difference seen in
https://github.com/mozilla/neqo/issues/2538.

Fixed in this commit.