fix(transport/qlog): call qlog::packet_io even without log::Level::Debug
https://github.com/mozilla/neqo/pull/2396 consolidated packet logging
and logging to qlog.  On a log level < `Debug`, it would return early,
i.e. not log the packet and not send it to qlog.

None of our other qlogs are bound to the current log level.

With this commit, all packets are sent to qlog, no matter the log level.