feat: make fast-apple-datapath optional, off by default (#2638)
The `quinn-udp` `fast-apple-datapath` Cargo feature allows Neqo to use
the private MacOS multi-message IO functions `sendmsg_x` and
`recvmsg_x`.

We have seen issues with `fast-apple-datapath`:

- https://github.com/quinn-rs/quinn/issues/2214
- https://github.com/quinn-rs/quinn/pull/2154

To reduce the risk of the upcoming larger roll out of the Fast UDP for
Firefox project to Firefox Release, disable `fast-apple-datapath` for
now.  Enabling it can then be a separate project afterwards.