Start Encoder start position in Buffer
One may create an `Encoder` encoding into a non-empty buffer via
`Encoder::new_borrowed_vec`. Track the starting position of the
`Encoder` in order to return the `Encoder` length in `Encoder::len`, and
not the potentially larger buffer length.

Useful in https://github.com/mozilla/neqo/pull/2593 as `PacketBuilder`
depends on `Encoder::len` representing the encoded length and not the
buffer length.