test(udp): set emsgsize test socket to blocking
The `send_ignore_emsgsize` unit test first sends an oversized datagram,
expecting it to be ignored, then sends a normal datagram, expecting it
to be received.

In the second check, to avoid race conditions, set the receiving socket
to blocking before the call to `receiver.recv`.

Fixes https://github.com/mozilla/neqo/issues/2752