ci: Don't generate and upload `junit.xml` to CodeCov anymore
`nextest` was producing `junit.xml` as a side effect, and so uploading it for [CodeCov's "flaky tests" dashboard](https://app.codecov.io/gh/mozilla/neqo/tests) was simple. When we stopped using `nextest` in #2679, I switched us to `junit-test` which I thought was converting the existing `cargo test` output to produce a `junit.xml` for upload.

Turns out what it is actually doing is compiling and running the entire `cargo test` suite again, which takes minutes in CI.

We're not really getting much use out of the CodeCov tests feature, so let's just stop feeding it data.