Improve resending of Initial on undecryptable
The goal here was to make the test more robust, but it turns out that
the code could have been simpler as well.  That is, rather than testing
for a separate threshold, we can test instead for whether the saved
datagram store is full.  That was the same constant anyway.

The test is made more robust by having it disable MLKEM, which ensures
that there is just one Initial packet in each flight.  The undecryptable
packet is constructed manually (by rewriting an Initial).