coalesce_05rtt robustness improvements (#2867)
* compatible_upgrade_large_initial robustness

This strips padding from all the packets in the handshake.
That makes all the accounting a lot simpler and robust.

This is not generic, but if I find a use for it in other tests (which
seems likely), it should be easy enough to factor it out.  The padding
test is a little fragile, but it accounts for the possibility of adding
SCONE at some point.

* Move the strip_padding function out

* Make the coalesce_05rtt test robust

This was checking for dropped packets as if that made a significant
difference.  It does, because we want to make sure that no packet was
dropped because it couldn't be decrypted, but we need to ensure that the
packet padding wasn't masking real drops.  The way to fix that is to
strip padding where necessary.  Factors out the code from the previous
test.

---------

Signed-off-by: Martin Thomson <mt@lowentropy.net>