Completely rewrite dup_server_flight1 (#2868)
* compatible_upgrade_large_initial robustness

This strips padding from all the packets in the handshake.
That makes all the accounting a lot simpler and robust.

This is not generic, but if I find a use for it in other tests (which
seems likely), it should be easy enough to factor it out.  The padding
test is a little fragile, but it accounts for the possibility of adding
SCONE at some point.

* Move the strip_padding function out

* Completely rewrite dup_server_flight1

This was very old and it showed.  The new version is much more in line
with other tests we've written recently.  It also avoids a bunch of
unnecessary back and forth.  It covers all the cases properly as well.

---------

Signed-off-by: Martin Thomson <mt@lowentropy.net>