Fix compatible upgrade
This changeset should have zero obvious effect right now, but this is
preparatory for the change that will randomize the first packet number.

This makes the whole version negotiation thing far more robust.  It
ensures that we correctly track the received packet numbers across
multiple crypto states at both client and server.

It changes how we confirm a compatible version upgrade, using the
availability of Handshake keys as the definitive trigger.  At the
server, that means that the server has processed the ClientHello (which
contains the client transport parameters and that is what the server
needs.  At the client, that means that the client has received at least
one packet with the chosen value, which is what the client needs (any
packet that makes progress on the handshake is enough).

This fixes two bugs:

1. We previously committed a version when receiving any CRYPTO frame.
   If we received an Initial packet with a TLS HelloRetryRequest, that
   would have been wrong.  That can happen if the server hadn't
   processed the ClientHello, but we were previously assuming that the
   server would have done so.

2. There are cases where a peer can send a v2 packet with a
   truncated packet number.  They make the choice to truncate based on
   the largest acknowledged, which might be > 256 if they are doing
   first packet number randomization.  Because we track the largest
   received packet number in the cryptographic send/receive state AND
   there is separate state for that for each version that might be used,
   if that truncated packet number is the first v2 packet we see,
   we fail to decrypt.  That situation is not recoverable.

There are now tests for these bugs, which are a little convoluted to
setup, but not completely implausible.

I've retained the case for compatible upgrade if the packet version
in an Initial packet changes during the handshake.  This is consistent
with what the spec requires and it can avoid the need to do the fiddly
stuff that follows.

Fixing the packet number continuation thing was easy on the server: it
is the one that decides to use a new version, so it can commit before it
ever needs to decrypt anything.  On the client, we have to be a little
more careful.  There's a tiny performance hit for doing the update,
which entails looking at the crypto state for every version, but it
should be modest overall.

There was one other test that tests an error condition that becomes
State::Closed rather than State::Closing as a result of deferring the
version confirmation.