PTO across packet number spaces better (#2878)
This tweaks the change that @larseggert made to reduce the number of
pointless handshake packets (Initial and Handshake both).  Rather than
cap the number of probes we send on a PTO, this keeps the limit of two
packets, but disables the probing we do with PING frames once we have
sent a packet of the indicated type.

This merges the spaces that we probe when a PTO occurs.  So we don't
just probe the last space that fired the PTO timer.   Because we time
out Initial and Handshake packets at the same time often, that results
in only probing the Handshake space.  That's probably OK in most cases,
but it can mean stalling Handshake if there is stuff to send at that
level at a server.

Because this allows more than one packet to be sent in Initial and
Handshake packet number spaces, there are some tests that do precise
counting and need to be adjusted.