Test and fix HRR bug (#2884)
I'm seemingly on a roll here.  I was looking to improve coverage, but I
found a bug instead.  We were not able to do a compatible upgrade if we
got HelloRetryRequest, because we were processing the transport
parameters twice.

Now, it's OK to process parameters twice, but not if you crash the
second time, which is what happened.

(We could check that they are the same, because they are supposed to be,
but that's not worth the extra complexity given the likelihood that
we'll ever see HelloRetryRequest in practice.)