feat(transport): randomize the first packet number (#2885)
* Randomize the first packet number

This is mostly the work of @larseggert, I'm just repackaging it.  This
will target his branch with any changes.

This randomizes the starting packet number the client uses for the Initial packet number space.

We don't randomize this on the server, since otherwise we'd need even more changes to the tests to account for that.

Fixes #2462.
Closes #2499.

* Use a stepped distribution

* Handle two byte packet numbers in the fuzzer

* Revert ACK rate changes to min_bandwidth

---------

Co-authored-by: Max Inden <mail@max-inden.de>