Add ECN marking to the simulator
This is still pretty crude, but it seems to approximate something
sensible now.

There was a big bug in the simulator where the TailDrop node wasn't
asking to be serviced when removing things from its queue.  So it was
only removing things from the queue when packets left the link.  That
meant that the only sensible way to use it was to have zero delay (as it
was being used).

This fixes that and adds ECN marking at a fixed threshold.  My
observation is that the marking threshold has a small impact on
performance (the output below shows that moving from marking at 0.7 to
marking at 0.8 improves throughput a little.  The first two were run at
0.7. Obviously, this comes at the cost of extra latency.  I have no real
idea what buffers make sense, this is just a synthetic test that sets
out some guard rails for things like CC and flow control under very,
very clean conditions.

This is still not ideal.  The min_bandwidth test should probably get a
higher throughput than it does.  Occasionally, this gets as high as
870Mbps.  But mostly it is 760Mbps or thereabouts.  I have no idea what
makes those occasional runs so much faster.

```
5.384 INFO Achieved 756.5999835520076 Mb/s bandwidth (link rate 1000)
5.411 INFO Achieved 759.7709198759945 Mb/s bandwidth (link rate 1000)
5.536 INFO Achieved 769.6899939983132 Mb/s bandwidth (link rate 1000)
5.358 INFO Achieved 771.442097560842 Mb/s bandwidth (link rate 1000)
5.423 INFO Achieved 768.9489036615845 Mb/s bandwidth (link rate 1000)
5.551 INFO Achieved 870.1786336328342 Mb/s bandwidth (link rate 1000)
5.504 INFO Achieved 768.9484584413776 Mb/s bandwidth (link rate 1000)
5.380 INFO Achieved 768.9018540063886 Mb/s bandwidth (link rate 1000)
5.412 INFO Achieved 768.9489036615845 Mb/s bandwidth (link rate 1000)
5.492 INFO Achieved 768.9481621100322 Mb/s bandwidth (link rate 1000)
5.418 INFO Achieved 872.2894963620984 Mb/s bandwidth (link rate 1000)
5.423 INFO Achieved 768.9489036615845 Mb/s bandwidth (link rate 1000)
5.377 INFO Achieved 872.2894966278358 Mb/s bandwidth (link rate 1000)
5.488 INFO Achieved 768.9024512780954 Mb/s bandwidth (link rate 1000)
5.439 INFO Achieved 767.510863851431 Mb/s bandwidth (link rate 1000)
5.487 INFO Achieved 768.9489038680874 Mb/s bandwidth (link rate 1000)
```