Make ECN RED useful
This adds some stats that show some very interesting results.

The most interesting result here is the buffer utilization metric I've
added as part of this.

A no-ECN run with a 4ms target buffer latency got 0.42051 utilization.
A run with ECN enabled uses a buffer that is twice the size.  It uses
the cubic curve that runs from 0 to 95% mark rate between 30 and 80% of
the buffer utilization ends up with 0.13187 utilization.
Those equate to no-ECN averaging 1682024 bytes in the buffer
and ECN averaging 1054930 bytes in the buffer.

Because the ECN does have a larger buffer, the maximum utilization in
bytes is 7744121 against the non-ECN 3999886 which completely fills up
in operation.  ECN loses no packets, no-ECN loses 6700.

The adjustment that doubles the buffer for the ECN-enabled case allows
throughput to be fairly similar. Without that, performance tanks badly.
This makes sense in a way: ECN means that we can't really use most of
the buffer (see above) so we lose the throughput benefits of buffering.

I suspected that we're hitting persistent congestion, but the log
statement there never shows in logs.  So it's not that.

What I see in the logs is that we get more congestion events with ECN
disabled.  And any we get are at a smaller congestion window.

There's a bunch of variability in the tests, even with a fixed seed for
the simulator.  It might pay to have a different configuration that
doesn't vary the length of connection IDs for some tests.  We can even
disable connection IDs.  That's the only thing I can think of that might
affect these tests.