Add ECN to the simulator (#2914)
* Add ECN marking to TailDrop

* Add ECN marking to the simulator

This is still pretty crude, but it seems to approximate something
sensible now.

There was a big bug in the simulator where the TailDrop node wasn't
asking to be serviced when removing things from its queue.  So it was
only removing things from the queue when packets left the link.  That
meant that the only sensible way to use it was to have zero delay (as it
was being used).

This fixes that and adds ECN marking at a fixed threshold.  My
observation is that the marking threshold has a small impact on
performance (the output below shows that moving from marking at 0.7 to
marking at 0.8 improves throughput a little.  The first two were run at
0.7. Obviously, this comes at the cost of extra latency.  I have no real
idea what buffers make sense, this is just a synthetic test that sets
out some guard rails for things like CC and flow control under very,
very clean conditions.

This is still not ideal.  The min_bandwidth test should probably get a
higher throughput than it does.  Occasionally, this gets as high as
870Mbps.  But mostly it is 760Mbps or thereabouts.  I have no idea what
makes those occasional runs so much faster.

```
5.384 INFO Achieved 756.5999835520076 Mb/s bandwidth (link rate 1000)
5.411 INFO Achieved 759.7709198759945 Mb/s bandwidth (link rate 1000)
5.536 INFO Achieved 769.6899939983132 Mb/s bandwidth (link rate 1000)
5.358 INFO Achieved 771.442097560842 Mb/s bandwidth (link rate 1000)
5.423 INFO Achieved 768.9489036615845 Mb/s bandwidth (link rate 1000)
5.551 INFO Achieved 870.1786336328342 Mb/s bandwidth (link rate 1000)
5.504 INFO Achieved 768.9484584413776 Mb/s bandwidth (link rate 1000)
5.380 INFO Achieved 768.9018540063886 Mb/s bandwidth (link rate 1000)
5.412 INFO Achieved 768.9489036615845 Mb/s bandwidth (link rate 1000)
5.492 INFO Achieved 768.9481621100322 Mb/s bandwidth (link rate 1000)
5.418 INFO Achieved 872.2894963620984 Mb/s bandwidth (link rate 1000)
5.423 INFO Achieved 768.9489036615845 Mb/s bandwidth (link rate 1000)
5.377 INFO Achieved 872.2894966278358 Mb/s bandwidth (link rate 1000)
5.488 INFO Achieved 768.9024512780954 Mb/s bandwidth (link rate 1000)
5.439 INFO Achieved 767.510863851431 Mb/s bandwidth (link rate 1000)
5.487 INFO Achieved 768.9489038680874 Mb/s bandwidth (link rate 1000)
```

* Easy to miss clippy on these

* Implement a basic RED scheme

* Remove lint expectations

* Change type to avoid cast

* Test that the distribution is approximately OK

* Make ECN RED useful

This adds some stats that show some very interesting results.

The most interesting result here is the buffer utilization metric I've
added as part of this.

A no-ECN run with a 4ms target buffer latency got 0.42051 utilization.
A run with ECN enabled uses a buffer that is twice the size.  It uses
the cubic curve that runs from 0 to 95% mark rate between 30 and 80% of
the buffer utilization ends up with 0.13187 utilization.
Those equate to no-ECN averaging 1682024 bytes in the buffer
and ECN averaging 1054930 bytes in the buffer.

Because the ECN does have a larger buffer, the maximum utilization in
bytes is 7744121 against the non-ECN 3999886 which completely fills up
in operation.  ECN loses no packets, no-ECN loses 6700.

The adjustment that doubles the buffer for the ECN-enabled case allows
throughput to be fairly similar. Without that, performance tanks badly.
This makes sense in a way: ECN means that we can't really use most of
the buffer (see above) so we lose the throughput benefits of buffering.

I suspected that we're hitting persistent congestion, but the log
statement there never shows in logs.  So it's not that.

What I see in the logs is that we get more congestion events with ECN
disabled.  And any we get are at a smaller congestion window.

There's a bunch of variability in the tests, even with a fixed seed for
the simulator.  It might pay to have a different configuration that
doesn't vary the length of connection IDs for some tests.  We can even
disable connection IDs.  That's the only thing I can think of that might
affect these tests.

* Fix comments, clean up 32-bit error

* Eliminate variability in benches

* no need to be public and trigger the missing_panics_doc lint

* Review comments

* Revise to quadratic and fix buffer sizes

* Fixup fuzzing

---------

Signed-off-by: Martin Thomson <mt@lowentropy.net>