refactor(http3): simplify RequestTarget
Previously to specify a request target we used the following types:

- trait `RequestTarget` for gettters
- struct `RefRequestTarget` as base representation of `RequestTarget`
- trait `AsRequestTarget` to convert other types into something that implements `RequestTarget`
- struct `UrlRequestTarget` implementing `RequestTarget` to convert a `str` potentially fallible

This commit reduces the mechanism to a single trait, i.e. `RequestTarget`.

Only downside is that one can no longer directly pass a `&str` to functions like
`fetch`. That said, we did this in a single place only, `neqo-bin`. Now, it
requires a single conversion step to a `Url` before. Additional upside is that
the caller handles the error on invalid request targets.