fix(transport): don't underflow on max MTU and GSO (#3000)
* test(transport): reach max MTU with GSO enabled

* fix(transport): don't underflow on max MTU and GSO

A GSO datagram batch is not allowed to exceed the MTU of the corresponding IP
address family. In the case of IPv4, this is `65535 - 20 - 8 = 65507`. In the case of
IPv6 this is `65535 - 40 -8 = 65487`.

Previously the GSO batch generation code has been conservative, assuming the
smaller IPv6 limit, i.e. `65487`. That said, on an IPv4 flow, PMTUD could
increase the MTU all the way to the IPv4 limit.

This was problematic here:

https://github.com/mozilla/neqo/blob/ae02833019335ad7dd19edd9f3ef7dc4b692b7c2/neqo-transport/src/connection/mod.rs#L2571

`DatagramBatch::MAX` is equal to the smaller IPv6 limit of `65487`.
`send_buffer.len` on an IPv4 flow with PMTUD can be up to the IPv4 limit, i.e.
`65507`. The subtraction can therefore underflow.

This commit fixes the issue, consulting the `PMTUD` module for the accurate
address family MTU instead.

* chore(Cargo.toml): prepare v0.16.1