feat: Use `jemalloc` as allocator
Gecko uses (it's own flavor of) `jemalloc`, including for Rust. The intend here is to get the standalone demo client and server, and the benchmarks, closer to Gecko.

As discussed with @valenting.

TODO: Is there a way to put this behind a feature or something, so that the `tikv-jemallocator` crate doesn't become part of the Gecko deps when vendored in?