fix(transport/packet): ensure sufficient space for packet length
`Builder::pn` adds the packet number and, iff it is a long packet, a dummy
packet length. Previously, it would only check for sufficient space for the
packet number, not the packet length. With a low packet size limit, this can
lead to later on building a packet longer than the limit. See
https://github.com/mozilla/neqo/issues/3046.

With this commit `Builder::pn` checks for sufficient space for both the packet
number and the packet length.