feat(transport): increase stream limits to 100
Allow a remote QUIC server to initiate up to 100 bidirectional and 100
unidirectional streams to the local client. In addition, decrease the connection
multiplier from 16 times the stream limit to 2 times the stream limit to reduce
the potential memory impact.

This aligns with Chromium's limits. A server can only initiate a stream when
using WebTransport. MOQ uses WebTransport. MOQ has the use-case to initiate
many concurrent unidirectional streams to the client.