perf(qlog): change to Option<Rc<RefCell<Option<SharedStreamer>>>>
Currently qlog is defined as:

```rust
pub struct Qlog {
    inner: Rc<RefCell<Option<SharedStreamer>>>,
}
```

https://github.com/mozilla/neqo/issues/1894 documents our finding, that
dereferencing the `Rc<RefCell<_>>` is costly, i.e. showing up in our CPU
profiles.

When writing to the qlog fails, we set the (inner) `Option` to `None`. Thus on
each consecutive logging atttempt we dereference the `Rc<RefCell<_>>` only to
find out that logging is disabled.

This patch still sets the inner `Option` to `None` to inform all other clones of
the `Rc` that logging is disabled. In addition, it introduces an outer `Option`,
which as well is set to `None`, thus avoiding the dereferencing of the
`Rc<RefCell<_>>` on each logging attempt once logging has failed.

---

Fixes https://github.com/mozilla/neqo/issues/1894.
Alternative for https://github.com/mozilla/neqo/pull/3005.