Don't collect when decoding a header block (#3197)
The code here used `collect()` just to check if the result was empty.

I've removed the debug assertions; I think we're good for that given the
age and stability of this code.  Happy to restore the check that the
required insert count matches, but there's no point in iterating over
the entire collection to determine if there are duplicates.  This code
is the only place that adds and there is no way we'll get duplicates.