fix: Stop loss monitoring for PMTUD triggering
The current method of retriggering PMTUD based on loss monitorig turns out to trigger very frequently on even somewhat lossy paths, such as some VPNs. This change disables loss monitoring as a trigger for PMTUD. Instead, we trigger PMTUD probing when we think the path has changed, i.e., on migration or when receiving a path challenge.

Also adds a test for the "VPN" case.